Add get_with_replicas(key, replicas)
Adds logic to be able to retrieve responsbile node for a key plus
the next `replicas` nodes after it. This is a common requirement for
implementations that use consistent hashing for p2p applications that
need to have some form of consistency. The `replicas` parameter is
typically referred to as replication factor in these implementations.

This function returns None if the ring is empty and will shrink
`replicas` if it is larger than the length of the ring to avoid
duplication.

Includes tests